fix(api): isolate MemFS proxy shell env helper
Keep Desktop MemFS git proxy shell environment injection independent from mocked backend API client modules so shell-env tests and command execution do not inherit stale test doubles.
