+++ src/agent/memoryGit.ts
@@ -25,11 +25,11 @@ import {
 import { homedir, platform } from "node:os";
 import { join } from "node:path";
 import { promisify } from "node:util";
+import { getClient } from "../backend/api/client";
 import {
-  getClient,
   getMemfsGitProxyRewriteConfig,
   getMemfsServerUrl,
-} from "../backend/api/client";
+} from "../backend/api/memfs-git-proxy";
 import { debugLog, debugWarn } from "../utils/debug";
 
 const execFile = promisify(execFileCb);
